# Conflicts:
#	src/main/java/dev/pro/shelter/listener/TelegramBotUpdatesListener.java
#	src/main/java/dev/pro/shelter/model/Cat.java
#	src/main/java/dev/pro/shelter/model/CatAdopter.java
#	src/main/java/dev/pro/shelter/model/Dog.java
#	src/main/java/dev/pro/shelter/model/DogAdopter.java
#	src/main/java/dev/pro/shelter/model/Pet.java
#	src/main/java/dev/pro/shelter/model/PetAdopter.java
#	src/main/java/dev/pro/shelter/model/Report.java
#	src/main/java/dev/pro/shelter/model/Users.java
#	src/main/java/dev/pro/shelter/service/CatService.java
#	src/main/java/dev/pro/shelter/service/DogService.java
#	src/main/java/dev/pro/shelter/service/impl/CatServiceImpl.java
#	src/main/java/dev/pro/shelter/service/impl/DogServiceImpl.java
#	src/main/resources/application-dev.properties
#	src/main/resources/application.properties
#	src/main/resources/db/changelog/report.sql
#	src/main/resources/db/changelog/users.sql

+++ src/main/java/dev/pro/shelter/service/impl/CatServiceImpl.java
@@ -32,7 +32,7 @@ public class CatServiceImpl implements CatService {
     }
 
     @Override
-    public Cat readCat(int id) {
+    public Cat readCat(long id) {
         logger.info("Вызван метод read с данными" + id);
         Optional<Cat> searchCat = catRepository.findById(id);
         if (searchCat.isEmpty()) {
@@ -55,7 +55,7 @@ public class CatServiceImpl implements CatService {
     }
 
     @Override
-    public Cat deleteCat(int id) {
+    public Cat deleteCat(long id) {
         logger.info("Вызван метод delete с данными" + id);
         Optional<Cat> searchCat = catRepository.findById(id);
         if (searchCat.isEmpty()) {
